% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_functions.R
\name{regionViews.delete}
\alias{regionViews.delete}
\title{Delete a resource view.}
\usage{
regionViews.delete(projectName, region, resourceViewName)
}
\arguments{
\item{projectName}{The project name of the resource view}

\item{region}{The region name of the resource view}

\item{resourceViewName}{The name of the resource view}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/}{Google Documentation}
}

