% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_functions.R
\name{regionViews.removeresources}
\alias{regionViews.removeresources}
\title{Remove resources from the view.}
\usage{
regionViews.removeresources(RegionViewsRemoveResourcesRequest, projectName,
  region, resourceViewName)
}
\arguments{
\item{RegionViewsRemoveResourcesRequest}{The \link{RegionViewsRemoveResourcesRequest} object to pass to this method}

\item{projectName}{The project name of the resource view}

\item{region}{The region name of the resource view}

\item{resourceViewName}{The name of the resource view}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/}{Google Documentation}

Other RegionViewsRemoveResourcesRequest functions: \code{\link{RegionViewsRemoveResourcesRequest}}
}

