% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{RegionViewsAddResourcesRequest}
\alias{RegionViewsAddResourcesRequest}
\title{RegionViewsAddResourcesRequest Object}
\usage{
RegionViewsAddResourcesRequest(resources = NULL)
}
\arguments{
\item{resources}{The list of resources to be added}
}
\value{
RegionViewsAddResourcesRequest object
}
\description{
RegionViewsAddResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to add resources to the resource view.
}
\seealso{
Other RegionViewsAddResourcesRequest functions: \code{\link{regionViews.addresources}}
}

