% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{RegionViewsInsertResponse}
\alias{RegionViewsInsertResponse}
\title{RegionViewsInsertResponse Object}
\usage{
RegionViewsInsertResponse(resource = NULL)
}
\arguments{
\item{resource}{The resource view object inserted}
}
\value{
RegionViewsInsertResponse object
}
\description{
RegionViewsInsertResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a resource view insert request.
}

