% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{RegionViewsListResourcesResponse}
\alias{RegionViewsListResourcesResponse}
\title{RegionViewsListResourcesResponse Object}
\usage{
RegionViewsListResourcesResponse(members = NULL, nextPageToken = NULL)
}
\arguments{
\item{members}{The resources in the view}

\item{nextPageToken}{A token used for pagination}
}
\value{
RegionViewsListResourcesResponse object
}
\description{
RegionViewsListResourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to the list resource request.
}

