% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{RegionViewsListResponse}
\alias{RegionViewsListResponse}
\title{RegionViewsListResponse Object}
\usage{
RegionViewsListResponse(nextPageToken = NULL, resourceViews = NULL)
}
\arguments{
\item{nextPageToken}{A token used for pagination}

\item{resourceViews}{The list of resource views that meet the criteria}
}
\value{
RegionViewsListResponse object
}
\description{
RegionViewsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to the list resource view request.
}
