% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{RegionViewsRemoveResourcesRequest}
\alias{RegionViewsRemoveResourcesRequest}
\title{RegionViewsRemoveResourcesRequest Object}
\usage{
RegionViewsRemoveResourcesRequest(resources = NULL)
}
\arguments{
\item{resources}{The list of resources to be removed}
}
\value{
RegionViewsRemoveResourcesRequest object
}
\description{
RegionViewsRemoveResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to remove resources from the resource view.
}
\seealso{
Other RegionViewsRemoveResourcesRequest functions: \code{\link{regionViews.removeresources}}
}

