% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ResourceView}
\alias{ResourceView}
\title{ResourceView Object}
\usage{
ResourceView(creationTime = NULL, description = NULL, id = NULL,
  labels = NULL, lastModified = NULL, members = NULL, name = NULL,
  numMembers = NULL, selfLink = NULL)
}
\arguments{
\item{creationTime}{The creation time of the resource view}

\item{description}{The detailed description of the resource view}

\item{id}{[Output Only] The ID of the resource view}

\item{labels}{The labels for events}

\item{lastModified}{The last modified time of the view}

\item{members}{A list of all resources in the resource view}

\item{name}{The name of the resource view}

\item{numMembers}{The total number of resources in the resource view}

\item{selfLink}{[Output Only] A self-link to the resource view}
}
\value{
ResourceView object
}
\description{
ResourceView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The resource view object.
}
\seealso{
Other ResourceView functions: \code{\link{regionViews.insert}},
  \code{\link{zoneViews.insert}}
}
