% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsAddResourcesRequest}
\alias{ZoneViewsAddResourcesRequest}
\title{ZoneViewsAddResourcesRequest Object}
\usage{
ZoneViewsAddResourcesRequest(resources = NULL)
}
\arguments{
\item{resources}{The list of resources to be added}
}
\value{
ZoneViewsAddResourcesRequest object
}
\description{
ZoneViewsAddResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to add resources to the resource view.
}
\seealso{
Other ZoneViewsAddResourcesRequest functions: \code{\link{zoneViews.addresources}}
}

