% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsInsertResponse}
\alias{ZoneViewsInsertResponse}
\title{ZoneViewsInsertResponse Object}
\usage{
ZoneViewsInsertResponse(resource = NULL)
}
\arguments{
\item{resource}{The resource view object that has been inserted}
}
\value{
ZoneViewsInsertResponse object
}
\description{
ZoneViewsInsertResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to an insert request.
}
