% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsListResourcesResponse}
\alias{ZoneViewsListResourcesResponse}
\title{ZoneViewsListResourcesResponse Object}
\usage{
ZoneViewsListResourcesResponse(members = NULL, nextPageToken = NULL)
}
\arguments{
\item{members}{The full URL of resources in the view}

\item{nextPageToken}{A token used for pagination}
}
\value{
ZoneViewsListResourcesResponse object
}
\description{
ZoneViewsListResourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a list resource request.
}

