% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsListResponse}
\alias{ZoneViewsListResponse}
\title{ZoneViewsListResponse Object}
\usage{
ZoneViewsListResponse(nextPageToken = NULL, resourceViews = NULL)
}
\arguments{
\item{nextPageToken}{A token used for pagination}

\item{resourceViews}{The result that contains all resource views that meet the criteria}
}
\value{
ZoneViewsListResponse object
}
\description{
ZoneViewsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a list request.
}
