% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsRemoveResourcesRequest}
\alias{ZoneViewsRemoveResourcesRequest}
\title{ZoneViewsRemoveResourcesRequest Object}
\usage{
ZoneViewsRemoveResourcesRequest(resources = NULL)
}
\arguments{
\item{resources}{The list of resources to be removed}
}
\value{
ZoneViewsRemoveResourcesRequest object
}
\description{
ZoneViewsRemoveResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to remove resources from the resource view.
}
\seealso{
Other ZoneViewsRemoveResourcesRequest functions: \code{\link{zoneViews.removeresources}}
}
